requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.27
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20
httpx>=0.27
pydantic>=2.6
pydantic-settings>=2.1
apscheduler>=3.10
rapidfuzz>=3.6
python-dotenv
typer>=0.9.0
prometheus-fastapi-instrumentator>=7.0
prometheus-client>=0.20
